# src/main/java/com/scheduler/controller/ScheduleRecordController.java
package com.scheduler.controller;

import com.scheduler.dto.ScheduleRecord.ScheduleRecordCreateDto;
import com.scheduler.dto.ScheduleRecord.ScheduleRecordResponseDto;
import com.scheduler.service.ScheduleRecordService;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;

import java.util.List;
import java.util.UUID;

import static org.springframework.http.ResponseEntity.ok;
import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
import org.springframework.stereotype.Controller;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

import javax.validation.Valid;

@RestController
@ConditionalOnProperty(
        value="controller.schedulerecord.active",
        havingValue = "true",
        matchIfMissing = true)
@RequestMapping("/api/v1/scheduleRecords")
@Validated
public class ScheduleRecordController {

    private final ScheduleRecordService scheduleRecordService;

    public ScheduleRecordController(ScheduleRecordService scheduleRecordService) {
        this.scheduleRecordService = scheduleRecordService;
    }

    @GetMapping("/all")
    public ResponseEntity<List<ScheduleRecordResponseDto>> findAll() {
        return ok(scheduleRecordService.findAll());
    }

    @GetMapping("/{id}")
    public ResponseEntity<ScheduleRecordResponseDto> findById(@PathVariable UUID id) {
        return ok(scheduleRecordService.findById(id));
    }

    @PostMapping("")
    public ResponseEntity<ScheduleRecordResponseDto> save(@Valid @RequestBody ScheduleRecordCreateDto scheduleRecordCreateDto) {
        return ok(scheduleRecordService.save(scheduleRecordCreateDto));
    }

    @PutMapping("/{id}")
    public ResponseEntity<ScheduleRecordResponseDto> update(@PathVariable UUID id,
                                                            @Valid @RequestBody ScheduleRecordCreateDto scheduleRecordCreateDto) {
        return ok(scheduleRecordService.update(id, scheduleRecordCreateDto));
    }

    @DeleteMapping("/{id}")
    public ResponseEntity<Void> deleteById(@PathVariable UUID id) {
        scheduleRecordService.deleteById(id);
        return ok().build();
    }
}